requirements.txt
asgiref==3.7.2
Django==4.2.8
exceptiongroup==1.2.0
importlib-metadata==7.0.1
iniconfig==2.0.0
Markdown==3.5.1
packaging==23.2
pluggy==1.3.0
pytest==7.4.3
pytest-django==4.7.0
python-dotenv==1.0.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
zipp==3.17.0
